server/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pdfplumber
python-dotenv
